+++ src/clues.ts
@@ -1,3 +1,21 @@
+export interface Rule {
+  heading: string;
+  docs: string;
+  suggestion: (value?: number | string) => string;
+  rationale: string;
+  source: string[];
+  ok: string;
+  notOk: string;
+  listen?: string;
+  rules?: string[];
+  fn?: (item: string, alt: string) => string | boolean;
+  id?: string;
+}
+
+interface Clue {
+  [rule: string]: Rule;
+}
+
 export const clues: Clue = {
   notOnlySpace: {
     heading: "Empty alt text",
